services/scam-detection/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
deep-translator
tqdm

# API, inference, OCR, MongoDB
fastapi
uvicorn
torch
transformers
Pillow
opencv-python
easyocr
pytesseract
python-multipart
pymongo
python-dotenv
motor
